Revert "Merge pull request #208 from ThomasRooney/master"
This reverts commit 6e9ce2e9b516940dc551847f188818563409e340, reversing
changes made to 4a8639ef5e352e5c26a6dbd3365ffd258befe665.

We are very fortunate that our continuous integration software did not pick up on this merge as the QA department apparently missed how this did not build AT ALL !!
This is absolutely unacceptable, and we will be taking swift action, as soon as we get over the very busy holiday period.