FizzStringReturner may not have copied all characters
I noticed that the String construtor was being called without specifying
the start offset and count. This is unacceptable. How do I know if it's
copying all the characters unless I can actually see it in the code. In
this change, you'll get the distinct air of superiority as you have now
begun dealing with low-level constructs such as arrays. As they say, rub
on some arrays, and watch the women flock.